Refactored client side html population for the calls groups_get_response and groups_info_response. Added new response group_numbers_response for simple incremention/decremention of group counters, added group_id to button id attribute in class_view. The members list in the group_view now updates the new members for those already in the group, and populates the whole list for those joining the group. The same concept applies to the groups_get button list.
If you want to use mocha to test this, change occurrences of 'ohs79ld1' in testing_no_loop.js to your desired class id.

TODO: Admin side updates so that it doesn't repopulate the whole list each time a student joins/leaves a group, and the addition of a add-group-response and delete-group-response instead of groups_get_response on admin group add/delete.